+++ src/components/RiderApprovalComponents/Rider_Approval_Table.jsx
@@ -1,10 +1,10 @@
 import React from "react";
 import {Button, Table } from "reactstrap";
-import "../assets/css/RiderApprovalDashboard.css";
-import Badge from "./Badge";
-import images1 from '../assets/image/carlo.jpg';
-import images2 from '../assets/image/cliff.jpg';
-import images3 from '../assets/image/cs3.png';
+import "../../assets/css/RiderApprovalDashboard.css";
+import Badge from "../../components/RiderApprovalComponents/Badge";
+import images1 from '../../assets/image/carlo.jpg';
+import images2 from '../../assets/image/cliff.jpg';
+import images3 from '../../assets/image/cs3.png';
